debug warnings in gcc-enbi

+++ src/dataArg.c
@@ -57,7 +57,7 @@ uint16_t content_getSize(uint8_t* self) {
 
 void content_setNext(uint8_t* self, uint8_t* next) {
     uint8_t* nextDir = self + content_nextOffset(self);
-    uint64_t pointerTemp = (uint64_t)next;
+    uint64_t pointerTemp = (long)next;
     for (uint32_t i = 0; i < sizeof(uint8_t*); i++) {
         // aboid \0
         nextDir[i] = pointerTemp;
@@ -104,7 +104,7 @@ uint8_t* content_init_hash(Hash nameHash,
         memset(contentDir, 0, size);
     }
 
-    uint64_t pointerTemp = (uint64_t)next;
+    uint64_t pointerTemp = (long)next;
     for (uint32_t i = 0; i < sizeof(uint8_t*); i++) {
         // aboid \0
         nextDir[i] = pointerTemp;
@@ -261,7 +261,7 @@ uint8_t* content_getNext(uint8_t* self) {
         pointerTemp = (pointerTemp << 8);
         pointerTemp += val;
     }
-    next = (uint8_t*)pointerTemp;
+    next = (uint8_t*)(long)pointerTemp;
     return next;
 }
 
@@ -309,7 +309,7 @@ float arg_getFloat(Arg* self) {
 }
 
 Arg* arg_setPtr(Arg* self, char* name, ArgType type, void* pointer) {
-    uint64_t pointerTemp = (uint64_t)pointer;
+    uint64_t pointerTemp = (long)pointer;
     uint8_t contentBuff[8];
     for (uint32_t i = 0; i < sizeof(uint8_t*); i++) {
         // aboid \0
@@ -350,7 +350,7 @@ void* arg_getPtr(Arg* self) {
         pointerTemp = (pointerTemp << 8);
         pointerTemp += val;
     }
-    pointer = (void*)pointerTemp;
+    pointer = (void*)(long)pointerTemp;
     return pointer;
 }
 char* arg_getStr(Arg* self) {
